+++ Books.js
@@ -4,6 +4,7 @@
  */
 
 const Books = {};
+const Validation = require('./Validation');
 
 /**
  * Autoriaus isleistos knygos irasymas i duombaze.
@@ -14,11 +15,19 @@ const Books = {};
  * @returns {Promise<string>} Tekstas, apibudinantis, koks autorius ir kurias metais isleido knyga.
  */
 Books.create = async (connection, authorId, bookName, bookReleaseYear) => {
-    sql = 'INSERT INTO books (author_id,book_name, release_year)\
+    if (!Validation.isValidName(bookName)) {
+        return console.error('ERROR: invalid string entry.')
+    }
+    if (!Validation.IDisValid(authorId)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'INSERT INTO books (author_id,book_name, release_year)\
     VALUES ("'+ authorId + '", "' + bookName + '", "' + bookReleaseYear + '")';
-    [rows] = await connection.execute(sql);
-    const createBook = `"${bookName}" written by author whose ID =  ${authorId} in year ${bookReleaseYear} buvo sekmingai irasyta i knygu sarasa`
-    return createBook
+        [rows] = await connection.execute(sql);
+        const createBook = `"${bookName}" written by author whose ID =  ${authorId} in year ${bookReleaseYear} buvo sekmingai irasyta i knygu sarasa`
+        return createBook
+    }
 }
 
 /**
@@ -45,10 +54,15 @@ Books.listAll = async (connection) => {
  * @returns {Promise<string>} Sarasas su knygu objektais.
  */
 Books.findByName = async (connection, bookName) => {
-    sql = 'SELECT * FROM `books` WHERE `book_name` = "' + bookName + '"';
-    [rows] = await connection.execute(sql);
-    const byBookName = `Book details by the name of "${bookName}": Author ID = ${rows[0].author_id} released in year ${rows[0].release_year}.`;
-    return byBookName;
+    if (!Validation.isValidName(bookName)) {
+        return console.error('ERROR: invalid string entry.')
+    }
+    else {
+        sql = 'SELECT * FROM `books` WHERE `book_name` = "' + bookName + '"';
+        [rows] = await connection.execute(sql);
+        const byBookName = `Book details by the name of "${bookName}": Author ID = ${rows[0].author_id} released in year ${rows[0].release_year}.`;
+        return byBookName;
+    }
 }
 /**
  * 
@@ -57,10 +71,15 @@ Books.findByName = async (connection, bookName) => {
  * @returns {Promise<string>} Tekstas, pateikiantis informacija apie knyga rasta pagal isleidimo metus. 
  */
 Books.findByYear = async (connection, bookReleaseYear) => {
-    sql = 'SELECT * FROM `books` WHERE `release_year` = "' + bookReleaseYear + '"';
-    [rows] = await connection.execute(sql);
-    const byBookYear = `Book details by the year ${bookReleaseYear}: Author ID = ${rows[0].author_id}, book name "${rows[0].book_name}".`;
-    return byBookYear;
+    if (!Validation.IDisValid(bookReleaseYear)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'SELECT * FROM `books` WHERE `release_year` = "' + bookReleaseYear + '"';
+        [rows] = await connection.execute(sql);
+        const byBookYear = `Book details by the year ${bookReleaseYear}: Author ID = ${rows[0].author_id}, book name "${rows[0].book_name}".`;
+        return byBookYear;
+    }
 }
 
 /**
@@ -72,10 +91,15 @@ Books.findByYear = async (connection, bookReleaseYear) => {
  * @returns {Promise<string>} Tekstas, pateikiantis atnaujinta informacija apie knyga rasta pagal knygos ID.
  */
 Books.updateById = async (connection, bookId, propertyName, propertyValue) => {
-    sql = 'UPDATE books SET ' + propertyName + ' = "' + propertyValue + '" WHERE books.id =' + bookId;
-    [rows] = await connection.execute(sql);
-    const updatedBookByID = `Book whose ID = ${bookId} property ${propertyName} has been updated to "${propertyValue}"`
-    return updatedBookByID
+    if (!Validation.isText(propertyName)) {
+        return console.error('ERROR: invalid string entry.')
+    }
+    else {
+        sql = 'UPDATE books SET ' + propertyName + ' = "' + propertyValue + '" WHERE books.id =' + bookId;
+        [rows] = await connection.execute(sql);
+        const updatedBookByID = `Book whose ID = ${bookId} property ${propertyName} has been updated to "${propertyValue}"`
+        return updatedBookByID
+    }
 }
 
 /**
@@ -86,10 +110,18 @@ Books.updateById = async (connection, bookId, propertyName, propertyValue) => {
  * @returns {Promise<string>} Tekstas, pateikiantis atnaujinta pavadinima knygos, kuri rasta pagal knygos ID.
  */
 Books.updateNameById = async (connection, bookId, bookName) => {
-    sql = 'UPDATE books SET book_name = "' + bookName + '" WHERE books.id =' + bookId;
-    [rows] = await connection.execute(sql);
-    const updatedBookNameByID = `Book, whose ID = ${bookId}, name has been updated to "${bookName}"`
-    return updatedBookNameByID
+    if (!Validation.isValidName(bookName)) {
+        return console.error('ERROR: invalid string entry.')
+    }
+    if (!Validation.IDisValid(bookId)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'UPDATE books SET book_name = "' + bookName + '" WHERE books.id =' + bookId;
+        [rows] = await connection.execute(sql);
+        const updatedBookNameByID = `Book, whose ID = ${bookId}, name has been updated to "${bookName}"`
+        return updatedBookNameByID
+    }
 }
 
 /**
@@ -100,10 +132,15 @@ Books.updateNameById = async (connection, bookId, bookName) => {
  * @returns {Promise<string>} Tekstas, pateikiantis atnaujintus knygos isleidimo metus, kuri rasta pagal knygos ID.
  */
 Books.updateYearById = async (connection, bookId, bookReleaseYear) => {
-    sql = 'UPDATE books SET release_year = "' + bookReleaseYear + '" WHERE books.id =' + bookId;
-    [rows] = await connection.execute(sql);
-    const updatedBookYearByID = `Book, whose ID = ${bookId}, release year has been updated to "${bookReleaseYear}"`
-    return updatedBookYearByID
+    if (!Validation.IDisValid(bookId || bookReleaseYear)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'UPDATE books SET release_year = "' + bookReleaseYear + '" WHERE books.id =' + bookId;
+        [rows] = await connection.execute(sql);
+        const updatedBookYearByID = `Book, whose ID = ${bookId}, release year has been updated to "${bookReleaseYear}"`
+        return updatedBookYearByID
+    }
 }
 
 /**
@@ -113,10 +150,15 @@ Books.updateYearById = async (connection, bookId, bookReleaseYear) => {
  * @returns {Promise<string>} Tekstas pasakantis kuri knyga, pagal knygos ID, buvo istrinta.
  */
 Books.delete = async (connection, bookId) => {
-    sql = 'DELETE FROM books WHERE books.id =' + bookId;
-    [rows] = await connection.execute(sql);
-    const deletedBook = `Book whose ID = ${bookId} has been deleted from books list.`
-    return deletedBook
+    if (!Validation.IDisValid(bookId)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'DELETE FROM books WHERE books.id =' + bookId;
+        [rows] = await connection.execute(sql);
+        const deletedBook = `Book whose ID = ${bookId} has been deleted from books list.`
+        return deletedBook
+    }
 }
 
 /**
@@ -126,10 +168,15 @@ Books.delete = async (connection, bookId) => {
  * @returns {Promise<string>} Tekstas pasakantis kuri knyga, pagal autoriaus ID, buvo istrinta.
  */
 Books.deleteAllByAuthorId = async (connection, authorId) => {
-    sql = 'DELETE FROM books WHERE author_id =' + authorId;
-    [rows] = await connection.execute(sql);
-    const deletedBookByAuthor = `Book whose author ID = ${authorId} has been deleted from books list.`
-    return deletedBookByAuthor
+    if (!Validation.IDisValid(authorId)) {
+        return ('ERROR: not a valid entry.')
+    }
+    else {
+        sql = 'DELETE FROM books WHERE author_id =' + authorId;
+        [rows] = await connection.execute(sql);
+        const deletedBookByAuthor = `Book whose author ID = ${authorId} has been deleted from books list.`
+        return deletedBookByAuthor
+    }
 }
 
 module.exports = Books;
